wormchain: move governance consts to sdk
