+++ wormchain/x/wormhole/keeper/msg_server_wasmd.go
@@ -12,14 +12,6 @@ import (
 	"golang.org/x/crypto/sha3"
 )
 
-// WasmdModule is the identifier of the Wasmd module (which is used for governance messages)
-var WasmdModule = [32]byte{00, 00, 00, 00, 00, 00, 00, 00, 00, 00, 00, 00, 00, 00, 00, 00, 00, 00, 00, 00, 00, 0x57, 0x61, 0x73, 0x6D, 0x64, 0x4D, 0x6F, 0x64, 0x75, 0x6C, 0x65}
-
-var (
-	ActionStoreCode           GovernanceAction = 1
-	ActionInstantiateContract GovernanceAction = 2
-)
-
 // Simple wrapper of x/wasmd StoreCode that requires a VAA
 func (k msgServer) StoreCode(goCtx context.Context, msg *types.MsgStoreCode) (*types.MsgStoreCodeResponse, error) {
 	if !k.setWasmd {
@@ -33,12 +25,12 @@ func (k msgServer) StoreCode(goCtx context.Context, msg *types.MsgStoreCode) (*t
 		return nil, err
 	}
 	// Verify VAA
-	action, payload, err := k.VerifyGovernanceVAA(ctx, v, WasmdModule)
+	action, payload, err := k.VerifyGovernanceVAA(ctx, v, vaa.WasmdModule)
 	if err != nil {
 		return nil, err
 	}
 
-	if GovernanceAction(action) != ActionStoreCode {
+	if vaa.GovernanceAction(action) != vaa.ActionStoreCode {
 		return nil, types.ErrUnknownGovernanceAction
 	}
 
@@ -83,12 +75,12 @@ func (k msgServer) InstantiateContract(goCtx context.Context, msg *types.MsgInst
 	}
 
 	// Verify VAA
-	action, payload, err := k.VerifyGovernanceVAA(ctx, v, WasmdModule)
+	action, payload, err := k.VerifyGovernanceVAA(ctx, v, vaa.WasmdModule)
 	if err != nil {
 		return nil, err
 	}
 
-	if GovernanceAction(action) != ActionInstantiateContract {
+	if vaa.GovernanceAction(action) != vaa.ActionInstantiateContract {
 		return nil, types.ErrUnknownGovernanceAction
 	}
